% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g.js}
\alias{g.js}
\title{Execute JavaScript code.}
\usage{
g.js(code, echo = TRUE)
}
\description{
Execute JavaScript code.
}
\examples{
g.js("1 < 2")
}
\seealso{
Other execution functions: \code{\link{g.rb}},
  \code{\link{g.r}}, \code{\link{g}}
}
