% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g.r}
\alias{g.r}
\title{Execute R code.}
\usage{
g.r(code, echo = TRUE)
}
\description{
Execute R code.
}
\examples{
g.r("runif(10^3)")
}
\seealso{
Other execution functions: \code{\link{g.js}},
  \code{\link{g.rb}}, \code{\link{g}}
}
