% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g.rb}
\alias{g.rb}
\title{Execute Ruby code.}
\usage{
g.rb(code, echo = TRUE)
}
\description{
Execute Ruby code.
}
\examples{
g.rb("1 < 2")
}
\seealso{
Other execution functions: \code{\link{g.js}},
  \code{\link{g.r}}, \code{\link{g}}
}
