% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gget.js}
\alias{gget.js}
\title{Retrieve the variable defined in GraalVM JS.}
\usage{
gget.js(var)
}
\description{
Retrieve the variable defined in GraalVM JS.
}
\seealso{
Other paired variables: \code{\link{gget.rb}},
  \code{\link{gget.r}}, \code{\link{gget}},
  \code{\link{gset.js}}, \code{\link{gset.rb}},
  \code{\link{gset.r}}, \code{\link{gset}}
}
