% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gget.r}
\alias{gget.r}
\title{Retrieve the variable defined in GraalVM FastR.}
\usage{
gget.r(var)
}
\description{
Retrieve the variable defined in GraalVM FastR.
}
\seealso{
Other paired variables: \code{\link{gget.js}},
  \code{\link{gget.rb}}, \code{\link{gget}},
  \code{\link{gset.js}}, \code{\link{gset.rb}},
  \code{\link{gset.r}}, \code{\link{gset}}
}
