% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gget.rb}
\alias{gget.rb}
\title{Retrieve the variable defined in GraalVM Ruby.}
\usage{
gget.rb(var)
}
\description{
Retrieve the variable defined in GraalVM Ruby.
}
\seealso{
Other paired variables: \code{\link{gget.js}},
  \code{\link{gget.r}}, \code{\link{gget}},
  \code{\link{gset.js}}, \code{\link{gset.rb}},
  \code{\link{gset.r}}, \code{\link{gset}}
}
