% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gget}
\alias{gget}
\title{Retrieve the variable defined in a GraalVM language. The local variable must have been 
initialized by one of the language specific gget.* or ggset.* functions.}
\usage{
gget(var)
}
\description{
Retrieve the variable defined in a GraalVM language. The local variable must have been 
initialized by one of the language specific gget.* or ggset.* functions.
}
\seealso{
Other paired variables: \code{\link{gget.js}},
  \code{\link{gget.rb}}, \code{\link{gget.r}},
  \code{\link{gset.js}}, \code{\link{gset.rb}},
  \code{\link{gset.r}}, \code{\link{gset}}
}
