% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{graalvm.start}
\alias{graalvm.start}
\title{Start the GraalVM agent. The agent is normally started automatically upon the first
code ecxecution.}
\usage{
graalvm.start()
}
\description{
Start the GraalVM agent. The agent is normally started automatically upon the first
code ecxecution.
}
