% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gset.js}
\alias{gset.js}
\title{Assign the value to the paired variable in Graal JS and locally.}
\usage{
gset.js(var, value)
}
\description{
Assign the value to the paired variable in Graal JS and locally.
}
\seealso{
Other paired variables: \code{\link{gget.js}},
  \code{\link{gget.rb}}, \code{\link{gget.r}},
  \code{\link{gget}}, \code{\link{gset.rb}},
  \code{\link{gset.r}}, \code{\link{gset}}
}
