% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gset.r}
\alias{gset.r}
\title{Assign the value to the paired variable in Graal FastR and locally.}
\usage{
gset.r(var, value)
}
\description{
Assign the value to the paired variable in Graal FastR and locally.
}
\seealso{
Other paired variables: \code{\link{gget.js}},
  \code{\link{gget.rb}}, \code{\link{gget.r}},
  \code{\link{gget}}, \code{\link{gset.js}},
  \code{\link{gset.rb}}, \code{\link{gset}}
}
