% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gset.rb}
\alias{gset.rb}
\title{Assign the value to the paired variable in Graal Ruby and locally.}
\usage{
gset.rb(var, value)
}
\description{
Assign the value to the paired variable in Graal Ruby and locally.
}
\seealso{
Other paired variables: \code{\link{gget.js}},
  \code{\link{gget.rb}}, \code{\link{gget.r}},
  \code{\link{gget}}, \code{\link{gset.js}},
  \code{\link{gset.r}}, \code{\link{gset}}
}
