% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{graalvm.setup}
\alias{graalvm.setup}
\title{Set up the GraalVM agent}
\usage{
graalvm.setup(home, host = "localhost", port = 9876, rlibs = paste0(home,
  "/language/R/library"))
}
\arguments{
\item{home}{The home folder of the GraalVM installation}

\item{host}{The local host name at which the GraalVM agent is listening}

\item{port}{The port at which the GraalVM agent is listening}

\item{rlibs}{The value of the FastR R_LIBS environmental variable. The default 
value is calculated as paste0(graalvm.home, "/language/R/library").}
}
\description{
Set up the GraalVM agent
}
\examples{
graalvm.setup("~/work/graalvm-0.21")
}
